refactor(itemValidation): move validation away from db
