+++ src/db/piece.ts
@@ -1,7 +1,6 @@
 import { Note } from "../types/Note";
 import { Piece, PieceBase } from "../types/piece";
 import { getSeconds } from "../utils/time";
-import { CheckResult } from "./CheckResult";
 import { executeSql } from "./common";
 import { rowToNote, rowToPiece } from "./RowTransform";
 import { PieceRow } from "./RowTypes";
@@ -91,23 +90,6 @@ export const getNotificationId = async (id: number): Promise<number | null> =>
     await executeSql('SELECT notifId FROM Pieces WHERE id = ?', [id])
     // @ts-ignore
         .then((res) => res.rows._array[0].notifId);
-
-export const validatePiece = async (piece: Piece): Promise<CheckResult> => {
-    if (piece.name.length === 0) {
-        return Promise.resolve({ valid: false, errors: 'You forgot to enter piece title ✍' });
-    }
-
-    return await executeSql('SELECT COUNT(*) as count FROM Pieces WHERE name = ? AND NOT id = ?',
-        [piece.name, piece.id])
-        .then((res) => {
-            // @ts-ignore
-            if (res.rows._array[0].count > 0) {
-                return { valid: false, errors: 'Piece with the same name already exists' }
-            }
-
-            return { valid: true };
-        })
-};
 
 export const getPieceById = async (id: number): Promise<Piece | undefined> => {
     const piece = await fetchPiece(id);
