+++ src/store/appReducers.ts
@@ -1,6 +1,9 @@
-import { combineReducers } from "redux";
+import { applyMiddleware, combineReducers, createStore } from "redux";
+import thunk from "redux-thunk";
 import { pieces } from "./reducers/piece";
 import { plans } from "./reducers/plan";
 import { sessions } from "./reducers/session";
 
-export const practiceManagerApp = combineReducers({ plans, pieces, sessions });
+const practiceManagerApp = combineReducers({ plans, pieces, sessions });
+
+export const store = createStore(practiceManagerApp, applyMiddleware(thunk));
